new util module
